chore(td-30): convert bin/copy-binaries to TypeScript
The DoD claimed the 4 remaining `bin/` `.js` were all plugin fixtures under
`bin/plugins/available/**`. Only 3 are: `copy-binaries.js` is the second half
of `npm run build` — it copies into `dist/` what tsc does not emit, the
cluster's `.proto` definitions and the executable server entrypoint.

Converted, and run through `tsx` from the source tree:

    "build": "rm -Rf ./dist && tsc && npx tsx ./bin/copy-binaries.ts"

`tsx` (4.21.0) is already a devDependency and already how the repo runs its
other TypeScript tooling (`npx tsx .ci/scripts/prepare-coverage.ts` in CI).
Running the compiled `dist/bin/copy-binaries.js` instead was the other option
and was rejected: from `dist/bin/`, `path.join(__dirname, "..")` resolves to
`dist/`, so the script would look for its sources under `dist/lib/` and copy
into `dist/dist/` — the paths would have had to be re-based, on release
tooling whose failure mode is a published package silently missing its
`.proto` files.

Verified inside Docker (the host's node_modules carry a linux-arm64 esbuild,
so `tsx` cannot run here): `npm run build` produces
`dist/lib/cluster/protobuf/{command,sync}.proto` and
`dist/bin/start-kuzzle-server` with mode 755, and tsc still emits
`dist/bin/copy-binaries.js`, so `package.json`'s `files` entry stays valid.

js baseline 50 -> 49; `bin/` is down to its 3 plugin fixtures. Adopted into
strict (102 files). No spec: `sonar.sources` is `./lib`, so `bin/` is outside
the analysed and coverage-measured scope, and this is build tooling rather
than product code — the ADR's "a file with no spec ships one" rule targets
`lib/`.

Closes #2705
