+++ package.json
@@ -4,7 +4,7 @@
   "version": "2.56.0",
   "description": "Kuzzle is an open-source solution that handles all the data management through a secured API, with a large choice of protocols.",
   "scripts": {
-    "build": "rm -Rf ./dist && tsc && node ./bin/copy-binaries.js",
+    "build": "rm -Rf ./dist && tsc && npx tsx ./bin/copy-binaries.ts",
     "cucumber": "cucumber.js --fail-fast",
     "dev": "tsx watch start-kuzzle-dev.ts",
     "doc-error-codes": "node -r ts-node/register doc/build-error-codes",
